requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
python-dotenv